refactor: UserContext 파일 분리
